Create bash scripts to build/test project
We were getting an error since the swift package manager was bundling
our test dependencies in the executable build and could not find XCTest.
We solved that issue by bundling dependencies depending on a
GRIMOIRE_ISTESTING variable. Problem is you can't set the env variable
from a Makefile script since it spins a child process and sets variables
in that process, unaffecting the parent process we are working in. We
solve that issue by using make to run bash scripts in our parent shell.